Ver 2.0 Move
각 클라이언트가 접속을 해서 움직이는 것까지 구현
아직 클라이언트끼리 동기화는 되지 않은 상태

+++ Player.cpp
@@ -1,4 +1,5 @@
 #include "Player.h"
+#include <stdio.h>
 
 Player::Player(float pX, float pY, float pHP) {
 	real_X = pX;		real_Y = pY;
@@ -18,11 +19,11 @@ void Player::changeMove(int* moveState) {
 	move_State[1] = moveState[1];
 }
 
-void Player::update(float frame_time) {
+void Player::update(int frame_time) {
 	float v_PPS = (PLAYER_SPEED * KPH_TO_MPS * METTER_TO_PIXEL);
-	float v_PPMS = v_PPS * (frame_time / 1000.0f);
+	float v_PPMS = v_PPS * ((float)frame_time / 1000.0f);
 	if ((move_State[0] * move_State[1]) != 0)
 		v_PPMS /= 1.414;
-	real_X = move(PLAYER_SIZE, real_X, move_State[0] * v_PPMS, 2000 - PLAYER_SIZE);
-	real_Y = move(PLAYER_SIZE, real_Y, move_State[1] * v_PPMS, 2000 - PLAYER_SIZE);
+	real_X = move(PLAYER_SIZE, real_X, move_State[0] * v_PPMS, 2000.0f - PLAYER_SIZE);
+	real_Y = move(PLAYER_SIZE, real_Y, move_State[1] * v_PPMS, 2000.0f - PLAYER_SIZE);
 }
